Port the winit backend changes that landed on master to the winit 0.31 API:

 - The window event handling moved from EventLoopState to
   WinitWindowAdapter::dispatch_winit_window_event. The drag and drop
   handling and the pointer events of winit 0.31 moved along with it.
 - Touch ids are keyed on (DeviceId, FingerId) so the allocator keeps
   touches from different devices apart.
 - invoke_from_active_event_loop pushes to the event queue and wakes the
   proxy, since winit 0.31 has no typed user event.
 - The vello renderer creates an Arc<dyn Window>.
 - The femtovg WGPU renderer uses the asynchronous wgpu 30 init.
 - unstable-wgpu-30 is now available next to unstable-winit-031.

SlintEvent and the EventLoopBuilder alias are gone from slint::winit_031:
winit 0.31 has an untyped proxy and a non generic EventLoopBuilder, which
is reachable through the re-exported winit crate.

+++ api/cpp/tests/translator.cpp
@@ -22,7 +22,8 @@ struct TestTranslator : public slint::Translator
     slint::SharedString ntranslate(uint64_t n, std::string_view singular, std::string_view plural,
                                    std::string_view context) const override
     {
-        return slint::SharedString("ntranslate(n=") + slint::SharedString::from_number(n)
+        return slint::SharedString("ntranslate(n=")
+                + slint::SharedString::from_number(static_cast<double>(n))
                 + ", singular=" + singular + ", plural=" + plural + ", context=" + context + ")";
     }
 };
